\name{cods-package}
\alias{cods-package}
\alias{cods}
\docType{package}
\title{
\packageTitle{cods}
}
\description{
\packageDescription{cods}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cods}
\packageIndices{cods}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cods}

Maintainer: \packageMaintainer{cods}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
